Guess i changed alot of stuff, didnt realize it tho.
