--- 1.19/src/main/resources/META-INF/minecraft-provider.properties
@@ -1,1 +1,0 @@
-provider-name=net.yakclient.minecraft.bootstrapper.one_nineteen.Minecraft_1_19_Provider
