+++ cache/version-mappings.json
@@ -1,0 +1,4 @@
+{
+  "1.19.2": "net.yakclient.minecraft:minecraft-provider-default:1.0-SNAPSHOT"
+
+}
