+++ default/src/main/resources/META-INF/minecraft-provider.properties
@@ -1,0 +1,1 @@
+provider-name=net.yakclient.minecraft.provider._default.DefaultMinecraftProvider
