+++ settings.gradle.kts
@@ -1,2 +1,2 @@
 rootProject.name = "minecraft-bootstrapper"
-include("1.19")
+include("default")
