--- src/main/kotlin/net/yakclient/minecraft/bootstrapper/MinecraftAppInstance.kt
@@ -1,5 +1,0 @@
-package net.yakclient.minecraft.bootstrapper
-
-import net.yakclient.boot.AppInstance
-
-public interface MinecraftAppInstance : AppInstance
