+++ src/main/kotlin/net/yakclient/minecraft/bootstrapper/MinecraftBootstrapper.kt
@@ -1,30 +1,27 @@
 package net.yakclient.minecraft.bootstrapper
 
+import com.durganmcbroom.artifact.resolver.RepositorySettings
 import com.durganmcbroom.artifact.resolver.simple.maven.HashType
 import com.durganmcbroom.artifact.resolver.simple.maven.SimpleMavenArtifactRequest
-import com.durganmcbroom.artifact.resolver.simple.maven.SimpleMavenDescriptor
 import com.durganmcbroom.artifact.resolver.simple.maven.SimpleMavenRepositorySettings
-import kotlinx.cli.ArgParser
-import kotlinx.cli.ArgType
-import kotlinx.cli.required
-import net.yakclient.archives.ArchiveReference
-import net.yakclient.boot.AppInstance
-import net.yakclient.boot.Boot
-import net.yakclient.boot.BootApplication
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import net.yakclient.boot.component.ComponentContext
+import net.yakclient.boot.component.SoftwareComponent
+import net.yakclient.boot.dependency.DependencyGraph
 import net.yakclient.boot.maven.MavenDataAccess
 import net.yakclient.boot.store.CachingDataStore
-import net.yakclient.common.util.CAST
 import net.yakclient.common.util.immutableLateInit
 import net.yakclient.common.util.make
 import net.yakclient.common.util.resolve
-import net.yakclient.common.util.resource.ProvidedResource
 import java.io.File
 import java.io.FileOutputStream
-import java.lang.instrument.ClassDefinition
+import java.net.URL
 import java.nio.channels.Channels
 import java.nio.file.Files
 import java.nio.file.Path
-
+import java.util.logging.Level
+import java.util.logging.Logger
 
 private enum class MinecraftRepositoryType(
     val settingsProvider: (String) -> SimpleMavenRepositorySettings,
@@ -37,53 +34,56 @@ private enum class MinecraftRepositoryType(
     })
 }
 
-public class MinecraftBootstrapper : BootApplication {
-    private val updatedClasses: MutableMap<String, ByteArray> = HashMap()
-    private val instrumentation = net.bytebuddy.agent.ByteBuddyAgent.install()
-    internal var minecraftRef: MinecraftReference by immutableLateInit()
+public class MinecraftBootstrapper : SoftwareComponent {
+    private val logger = Logger.getLogger(this::class.simpleName)
+    public var minecraftHandler: MinecraftHandler<*> by immutableLateInit()
+        private set
 
     init {
         checkNotInitialized()
-
         instance = this
-        MinecraftMixinAccess.registerListener { name, bytes ->
-            updatedClasses[name] = bytes
-        }
     }
 
-    internal companion object {
-        var initialized: Boolean = false
-        var instance: MinecraftBootstrapper by immutableLateInit()
+    public companion object {
+        private var initialized: Boolean = false
+        public var instance: MinecraftBootstrapper by immutableLateInit()
 
-        fun checkNotInitialized() {
+        public fun checkNotInitialized() {
             check(!initialized) { "Minecraft already initialized" }
 
         }
 
-        fun checkInitialized() {
+        public fun checkInitialized() {
             check(initialized) { "Minecraft not initialized yet!" }
         }
     }
 
-    override fun newInstance(args: Array<String>): AppInstance {
-        val parser = ArgParser("minecraft-bootstrap", skipExtraArguments = true)
 
-        val version by parser.option(ArgType.String, "minecraft-version").required()
-        val gameProviderRepository by parser.option(ArgType.String, "game-provider-repo").required()
-        val gameProviderRepositoryType by parser.option(
-            ArgType.Choice<MinecraftRepositoryType>(),
-            "game-provider-repo-type"
-        ).required()
-        val gameCachePath by parser.option(ArgType.String, "game-cache-path").required()
 
-        parser.parse(args)
+    override fun onEnable(context: ComponentContext) {
+        logger.log(Level.INFO, "Minecraft Bootstrapper is enabling.")
+
+        val version by context.configuration
+        val repository by context.configuration
+        val repositoryType by context.configuration
+        val type = MinecraftRepositoryType.valueOf(repositoryType)
+        val cache by context.configuration
+        val providerVersionMappings by context.configuration
+        val mcArgs by context.configuration
+
+        val cachePath = Path.of(cache)
+
+        val providerVersionsPath = cachePath resolve "minecraft-versions.json"
+
+        val versionMappings = providerVersionsPath.toFile().takeIf(File::exists)
+            ?.let { ObjectMapper().readValue<Map<String, String>>(it) } ?: HashMap()
 
         val handler = MinecraftProviderHandler(
             { request, resource ->
                 val descriptor by request::descriptor
 
                 val jarName = "${descriptor.artifact}-${descriptor.version}.jar"
-                val jarPath = Path.of(gameCachePath) resolve descriptor.group.replace(
+                val jarPath = cachePath resolve descriptor.group.replace(
                     '.',
                     File.separatorChar
                 ) resolve descriptor.artifact resolve descriptor.version resolve jarName
@@ -99,47 +99,35 @@ public class MinecraftBootstrapper : BootApplication {
 
                 jarPath
             },
-            CachingDataStore(MavenDataAccess(Path.of(gameCachePath))),
-            Boot.maven,
+            CachingDataStore(MavenDataAccess(cachePath)),
+            context.bootContext.dependencyProviders["simple-maven"]?.graph  as DependencyGraph<SimpleMavenArtifactRequest, *, RepositorySettings>,
         ) {
-            val descriptor =
-                SimpleMavenDescriptor("net.yakclient.minecraft", "minecraft-provider-$it", "1.0-SNAPSHOT", null)
-
             SimpleMavenArtifactRequest(
-                descriptor,
+                versionMappings[it] ?: run {
+                    providerVersionsPath.make()
+
+                    Channels.newChannel(URL(providerVersionMappings).openStream()).use { cin ->
+                        FileOutputStream(providerVersionsPath.toFile()).use { fout: FileOutputStream ->
+                            fout.channel.transferFrom(cin, 0, Long.MAX_VALUE)
+                        }
+                    }
+
+                    ObjectMapper().readValue<Map<String, String>>(providerVersionsPath.toFile())
+                }[it] ?: throw IllegalArgumentException("Failed to find version provider for version '$it'"),
                 includeScopes = setOf("compile", "runtime", "import"),
             )
         }
 
-        val provider = handler.get(version, gameProviderRepositoryType.settingsProvider(gameProviderRepository))
+        val provider = handler.get(version, type.settingsProvider(repository))
 
-        minecraftRef = provider.getReference(version, Path.of(gameCachePath))
+        minecraftHandler = MinecraftHandler(
+            version,
+            cachePath,
+            provider,
+            mcArgs.split(';').toTypedArray()
+        )
+    }
 
-        updatedClasses.forEach {
-            val oldEntry = minecraftRef.archive.reader[it.key]!!
-            minecraftRef.archive.writer.put(
-                ArchiveReference.Entry(
-                    it.key,
-                    ProvidedResource(
-                        oldEntry.resource.uri
-                    ) { it.value },
-                    oldEntry.isDirectory,
-                    oldEntry.handle
-                )
-            )
-        }
-        @Suppress(CAST)
-        val get = (provider as MinecraftProvider<MinecraftReference>).get(minecraftRef)
-
-        MinecraftMixinAccess.registerListener { name, bytes ->
-            instrumentation.redefineClasses(
-                ClassDefinition(
-                    get::class.java.classLoader.loadClass(name),
-                    bytes
-                )
-            )
-        }
-
-        return get
+    override fun onDisable() {
     }
 }
