+++ src/main/kotlin/net/yakclient/minecraft/bootstrapper/MinecraftHandle.kt
@@ -1,0 +1,9 @@
+package net.yakclient.minecraft.bootstrapper
+
+import net.yakclient.archives.ArchiveHandle
+
+public interface MinecraftHandle  {
+    public val archive: ArchiveHandle
+
+    public fun start(args: Array<String>)
+}
