+++ src/main/kotlin/net/yakclient/minecraft/bootstrapper/MinecraftProvider.kt
@@ -1,9 +1,10 @@
 package net.yakclient.minecraft.bootstrapper
 
+import net.yakclient.archives.ArchiveHandle
 import java.nio.file.Path
 
 public interface MinecraftProvider<T: MinecraftReference> {
-    public fun getReference(version: String, dumpPath: Path) : T
+    public fun getReference(version: String, cachePath: Path) : T
 
-    public fun get(ref: T) : MinecraftAppInstance
+    public fun get(ref: T) : MinecraftHandle
 }
