+++ src/main/kotlin/net/yakclient/minecraft/bootstrapper/MinecraftReference.kt
@@ -3,5 +3,6 @@ package net.yakclient.minecraft.bootstrapper
 import net.yakclient.archives.ArchiveReference
 
 public interface MinecraftReference {
+    public val version: String
     public val archive: ArchiveReference
 }
