--- src/main/resources/META-INF/app.json
@@ -1,15 +1,0 @@
-{
-  "name": "Minecraft",
-  "appClassName": "net.yakclient.minecraft.bootstrapper.MinecraftBootstrapper",
-  "dependencies": [
-    {
-      "type": "simple-maven",
-      "request": {
-        "descriptor": "net.bytebuddy:byte-buddy-agent:1.12.18"
-      },
-      "repository": {
-        "url": "https://repo.maven.apache.org/maven2"
-      }
-    }
-  ]
-}
