--- src/main/resources/META-INF/mixins/compliance.properties
@@ -1,1 +1,0 @@
-mixin-access-classname=
